Allow for individual user metrics to be optionally ignored
