+++ .goreleaser.yml
@@ -13,8 +13,8 @@ builds:
   - darwin
 dockers:
 - image_templates:
-  - "kumina/openvpn-exporter:latest"
-  - "kumina/openvpn-exporter:{{ .Tag }}"
+  - "kumina/openvpn-exporters:latest"
+  - "kumina/openvpn-exporters:{{ .Tag }}"
   build_flag_templates:
   - "--label=org.label-schema.schema-version=1.0"
   - "--label=org.label-schema.version={{ .Version }}"
